move mock iterator into mock dir
